:page_facing_up: Refactored output file functions (#45)
* :sparkles: Refactored countStatistics function

* :boom: Added helper function for calculating ratio

* :hammer: Fixed possible loss of data warning

* :sparkles: Shortened initializer list for game statistics

* :zap: Added ArmyStatistics struct to be used in statistics

* :pencil2: Added a missing line

+++ UnitClasses/EarthGunnery.cpp
@@ -77,5 +77,5 @@ bool EarthGunnery::attack()
 
 int EarthGunnery::getPriority() const
 {
-    return getHealth() + getPower();
+    return (int) getHealth() + getPower();
 }
